+++ CloneCraft/Generator/Noise/OctavePerlin.cpp
@@ -1,6 +1,6 @@
 #include "OctavePerlin.h"
 
-OctavePerlin::OctavePerlin(int octaves, float persistence, float frequency)
+OctavePerlin::OctavePerlin(int octaves, double persistence, double frequency)
 	: m_octaves{ octaves }, m_persistence{ persistence }, m_frequency{ frequency } {
 	for (int i = 0; i < m_octaves; ++i) {
 		noises.push_back(PerlinNoise());
@@ -11,17 +11,17 @@ OctavePerlin::~OctavePerlin() {
 }
 
 double OctavePerlin::getNoise(vec2 pos) const {
-	float total = 0;
-	float frequency = m_frequency;
-	float amplitude = 1;
-	float maxTotal = 0;
-	for (int i = 0; i < noises.size(); i++) {
-		total += noises[i].getNoise(pos * frequency) * amplitude;
+	double total = 0;
+	double frequency = m_frequency;
+	double amplitude = 1;
+	double maxTotal = 0;
+	for (int i = 0; i < int(noises.size()); i++) {
+		total += noises[i].getNoise(pos * static_cast<float>(frequency)) * amplitude;
 
 		maxTotal += amplitude;
 
 		amplitude *= m_persistence;
-		frequency *= 2;
+		frequency *= 2.;
 	}
 
 	return total / maxTotal;
